requirements.txt
langchain
langchain-openai
langchain-community
langchain-core
langchain-text-splitters
langgraph
python-dotenv
openai
beautifulsoup4
tiktoken 